+++ src/main/java/com/lalongooo/videocompressor/OnCompress.java
@@ -1,6 +1,6 @@
 package com.lalongooo.videocompressor;
 
 public interface OnCompress {
-    void onCompressProgress(String id, long percent);
+    void onCompressProgress(String id, int percent);
     void onCompressFinish(String id, boolean compress);
 }
